Removed unintentional semi-colon at the end of some if statements.
Henk - I'm 99% sure you didn't mean to Initialize() regardless of the result of Start{Push/Pull}Dialog() so I removed the semi-colon. Visual Studio was giving a warning about this and I wanted to clean up the project so it compiles without any warnings.

If this is not the case you can just revert this commit.