Uses InsertMenu() instead of InsertMenuItem() to inject Git Extensions submenu now.
For some reason InsertMenuItem was being inconsistent in injecting the submenu into the context menu.  On some users' machines the Git Extensions item would never show up when right clicking, on others it would only show up when right-clicking files and not directories.  Debugging showed that even though InsertMenuItem() returned true the item was not visible.  Using InsertMenu appears to fix this for all cases we had where it failed.