Show remote branches in "Contained in:" status line
There was a small issue when browsing commits on a branch without local
tracking branch set up. The revision grid would annotate the head commit
with a "[origin/branchname]" tag, but its status line read "Contained in: no branch",
the same applied to other commits in that branch.
The fix works by fetching and displaying remote branches, but only if there
are no local branches for selected commit. This is to prevent screen getting
cluttered with remote duplicates of local branches, without the overhead of
explicitly eliminating remote branches tracked by local ones.