Fixed merge scripts for use in Git Extensions
-Replaced sMyDoc for sMergedDoc, now the saved file is picked up by git (Open Office only)
-Removed TSVN from messagebox title

The Open Office merge scripts do work pretty good. The Microsoft Word scripts are a bit crappy, but it is better then nothing. The user needs to use the save-as button to save the file to the correct location. To help the users, a warning has been added when the file is not modified after running the merge script.