When there is only one repository host plugin loaded, use that description
When only the github plugin is loaded, the menus item is "Github" instead of "Rository host". This should make it easier for users to find the github integration. The menu item is hidden when there is no repository host.

I also changed the installer to make it possible to not install certain plugins. By default all plugins are enabled for now.