Added a method to create a CommitInformation object using the raw string output of a git command
The plan is to use this to show git notes in commit details in much the same way that gitk shows them when running with --show-notes.  It would have been much easier to do this using %N in the format like for git-show but unfortunately it did not appear as though I could display what gitnotes ref the notes were from.