Ignore and remove .resources file.
This appears to be generated during the post build step of the GitPlugin csproj, so every time a build gets run, this file shows up as a changed file in git.  From what I can tell since the file is being generated it appears we can remove it from source control and add it to .gitingore.