Removed the necessity to have a blank line at the top of the .git* files.
The original issue was the the BOM marker was being output to the files automatically by the StreamWriter.  This was fixed by avoiding the stream writer altogether and instead writing bytes directly to the file.
We avoid the issue by not writing the contents of `Encoding.GetPreamble()` to the file.

This fixes #638.