Fixed the Test Connection functionality of the Manage Remotes UI.
The Manage Remotes UI accepts a URL.  This is fine and that URL works when used by git.  However the Manage Remotes UI has a Test Connection Button.  This button invokes plink.exe directly, passing it the raw URL.  plink.exe doesn't take a URL as an argument, so if we have an ssh:// format URL we parse it into a form which works with plink.exe.