fixed an issue with empty commit templates showing watermark text
Prior to the patch, when using a commit template with blank contents for the text, the commit-message-box would change to "Enter commit message" as watermark, yet the box itself already being focused.

Thus you were able to edit the watermark text, and when unfocusing/refocusing the text again the whole commit message would disappear.

This patch addresses this, in that the message is explicitly focused and the watermark explicitly hidden.