Moving forms to .net 4.0 trying to understand still issue #1037
This is still experimental. Here we are trying to address .net 4.0, we are upgrading the resx so instead of using .net 2.0, let's use .net 4.0. I've been testing this for a while but I need help to test all functionality. It's too soon to say it's more stable or that this commit will fix the issues.

I did not test with mono, will do this soon.

Changing the form would eventually change the design code behind and changing some properties like if you change DcvsGraph.cs and RevisionGrid properties for some reason, VS will add code that will crash gitextensions.

Valid to note too is that we still have .net 1.0 and 2.0 DLL's dependencies that rely on 3rd party dll's like the CSharpEditor used in FormCommit, or Extensibility reference in GitPlugin project.

LAst but not least I had the crash behavior while creating the git process, class GitCommandHelpers.cs. The startInfo class does not implement IDisposable so before the return I'm assigning null to it.