Don't warn about pushing to a branch if it is the default for the remote
This avoids warning that a branch doesn't exist if it is the default branch for
the remote that is being pushed to.  Since the branch has been configured
as part of the remote, it is not likely to be a transient typo.

This is useful for workflows that involve creating push shortcuts for non-public
branches when using systems like Gerrit.