Added option to not use the commit dialog to specify the commit message
To better support git workflows which use standard git commit templates or
prepare-commit-msg hooks, the "Write Commit Messages in Commit Dialog" option
has been added to the gitextensions tab of the settings window.

When it is unchecked, the commit dialog won't specify a commit message to the
git-commit, allowing git-commit to do it's behaviour of launching $EDITOR giving
the user the ability to edit the auto-generated commit message, just like on the
command line.

By default the original behaviour is used. (Require the message to be entered
in the commit dialog)