Use a checkbox for Commit Amend
Introduces a check box for Commit Amend that replaces the previous
Amend button in the Commit dialog.

When the Amend Commit box is checked, it will load the commit
message of the previous commit, and the Commit button will
result in a "commit --amend".

Ideally in the future, when the box is checked, the staging area on the
the left of the Commit dialog can include changes from the previous
commit for easy amending.

This is intended to address issue #565