Change FormCommit to use TextChanged event instead of KeyPress
Soon, lines that are too long will be marked red for characters that cross
the configured boundary for max line length. For this to work properly, we
will need to use the TextChanged event. The reason is that the KeyPress
event is invoked *before* the key that was pressed has been assigned to the
TextBox.Text property.