Add red coloring to long subject line in commit message
If the config is set to allow max 72 chars in the commit message subject
line, then any character exceeding 72 chars will be colored red, so as to
indicate to the user that he should shorten it.

The patch also does some simple refactoring; the method is renamed to
FormatLine - now taking a zero-offset line parameter.

The if tests have also been slightly modified; instead of comparing the
limit to Message.CurrentColumn, the limit is now compared with the
lineLength. The reason is that a user might type a long line, then click
somewhere in the middle of the line and press backspace... The line may
still be longer than the allowed limit.