Add config option to enable/disable indentation in commit messages
This patch adds a checkbox to the commit template settings window, which
controls a newly added config CommitValidationIndentAfterFirstLine. By
default this is checked, so the behaviour is the same as before this patch.

If unchecked, then commit messages will not have a bullet " - " appended on
line 3 of the commit message, and no indent will be inserted in the
WordWrap method, when newlines are created.