Add event TextAssigned to EditNetSpell. Use to format commit msgs
This event is invoked when the Text property is set programatically, and we
use this to call the FormatAllText method. FormatAllText simply calls
FormatLine for all lines in the text, such that any formatting is applied
to text that is assigned to the commit message text box.

Useful when a cached message is assigned to Message.Text. It might need
coloring or word-wrapping.