Rename/Modify EditNetSpell.ForceNextLine to EnsureEmptyLine
This patch changes this method so that it doesn't matter if it is called at
the end of some text *or* for any line of existing text. If the latter is
the case, it will do nothing if the line of the supplied 'line' argument is
already an empty line. In other cases it will insert a new line.

If any insertion is done, it will not explicitly set the Text property, but
rather modify the SelectedText. This has the benefit that existing coloring
is preserved.