Add option to turn off automatic word-wrapping in commit messages
This patch adds a checkbox called Auto-wrap to the status bar. If checked
(which is the default), it will behave as before, so that text below the
subject line is automatically word-wrapped.

If unchecked, no wrapping is done, and instead line coloring is performed
as for the subject line. This is useful for people who might insert ascii
art into commit messages, and thus occasional breaching of max length for
lines might be acceptable. Also useful for users who simply want to do line
wrapping manually.