Always reset to black after changing some text color
If the cursor was placed at some section where the text color was already
set to red, and something triggered this method, we could in some situations
end up with a situation where the cursor moved to a new line, and *all* text
became red after that point.

This was not desired, and this method simply changes the behaviour so that
after a color change, the color at the cursor position will be set to black.