Improve how EditNetSpell.EnsureEmptyLine works
First of all, there is no need to do anything at all unless someone tries to
type anything on that line-that-should-be-empty.

Secondly, when someone types something at the second line, we want to
prepend whatever-that-was with a newline. So we insert the newline (and
optionally, the bullet) at the *beginning* of the line. Then we place the
cursor after the newline and whatever-the-user-typed.