Move word-wrapping logic to new class WordWrapper
This class has quite simple logic for wrapping text at a fixed limit. It
does this by the standard greedy algorithm where you keep adding words as
long as that doesn't cause the line to be longer than the limit.

There is an exception to that rule, and that is if someone were to enter a
single word on a line.. Then we will accept it no matter what the limit is.

This patch also separates the wrapping logic from the EditNetSpell control,
which seems like a good idea to me.