[Setup] Add choice to add install dir to PATH
Git Extensions' installation directory will be added to the system
environment variable "PATH" if selected. Defaults to yes.
Always add to the _system_ variable and not user variable, since
the install package can't install without administrative privileges
anyway, so the "install for this user only" option is broken anyway.

The directory will be removed from PATH when Git Extensions
is uninstalled.

Solves issue #438 (which is a feature request).