Refactor and decrease time of adding submodules to dropdown list
Every submodule has a new thread started to get the image for the
submodule dropdown list. This is fine for a small amount of submodules
but is very slow for a large number of submodules.

Benchmarking the time to finish populating all information for the
submodule dropdown list with 150 submodules takes about 10 seconds on
average. In addition, a large amount of the work is done on the UI
thread causing the application to not be responsive.

Refactored the code to not be in one giant function, all information
is retrieved on a background thread, and parallel threads are used to
get submodule status information. This gives about an 80% performance
increase for 150 submodules.

Other changes:
- Show a "Loading..." menu item while submodule information is being
   retrieved.
- Removed the GetCurrentSubmoduleLocalPath() function from the
   GitModule class as it was only used in the submodule dropdown code
   and does the same thing as the SubmodulePath property.
- Created a ToolStripItemInfo class to hold information for a
   ToolStripItem while on a background thread.