Fix search for commit hash or large number in commit messages
PR #1477 added ability to find a commit by SHA hash in quick filter.
If the value in quick filter looks like a SHA hash, Git Extensions removes
--grep command line option and adds _MessageFilter and _ShaFilter
to RevisionGridInMemFilter. RevisionGridInMemFilter filters revisions
with the required SHA hash and revisions that contain the SHA hash
in the commit message (GitRevision.Message). So, it should not have
break anything. But the problem is that GitRevision.Message contains
only commit subject, not the full commit message.

In the previous commit I renamed GitRevision.Message to
GitRevision.Subject to avoid such mistakes in the future.

In this commit I add filling of GitRevision.Body (it was null) and using it
for filtering in RevisionGridInMemFilter. It's removed after filtering to
leave memory consumption on the same level.

Closes #2840