Fix infinite loop bug during undo with auto-wrap
If Undo (Ctrl-Z) was used to undo a line that was created due to
auto-wrapping, the result was an infinite loop.

UndoHighlighting() invokes Undo() repeatedly, but if that undo removes
a line break introduced by auto-wrap, Message_TextChanged() runs
immediately and re-wraps it, causing UndoHighlighting() to continue
trying to undo forever.

This fixed the issue by not performing auto formatting during an undo
operation. This should not affect the resulting functionality  because
those undo operations should only restore text that has already
been auto-formatted anyway.