Stash:  Use auto stash setting instead of manual stash setting for browse/commit form.
The existing behavior was very difficult to understand; I was unable to learn
it without examining the Git Extensions source code myself.  Specifically, it's
a natural user expectation that if you check "Include untracked files in stash"
in the Settings dialog, this setting will be honored when choosing to make a
stash from the main Browse dialog.  This commit makes it so.