Add newtonsoft.Json version bindingRedirect for system.http.format
So system.http.format can use the newer version of newtonsoft.json.dll that is being used by gitextensions.
Otherwise the program won't be able to load the build server adapters properly and consequently fails to retrieve and show the build results in the revision grid.
Also fails to show the build server types in the dropdown list in the build server integration settings page.