Trying to rule out which output function writes to where.
For a starter, extract writing the output string ("log") so that writes to it were explicit. There are still same-named functions for writing either to display control only or to both places, that should better be renamed as well.
Also upd echo to skip it with conemu, it echoes on itself.