Force the TFS plugin to use new version of Json dll
One of hte the TFS SDK dlls is referencing the v4.5 Json dll; therefore will copy the v4.5 Json dll to the plugins dll.
Tried to make the TFS dlls private(CopyLocal to true), but doesn't help.
So I decided to add the new Json to the plugin's references, thus copying a new version to the plugins folder