Get value of status.showUntrackedFiles using Module.EffectiveConfigFile
... instead of Module.RunGitCommand, as Arkady suggested.

Note this is prone to some error.  If the .git/config file contains:

<<<
[status]
[status]
      showUntrackedFiles = no
<<<

, then Module.EffectiveConfigFile.GetValue("status.showUntrackedFiles") wont' return the correct value,
but the one found in in ~/.gitconfig.

It looks as the mechanism used by EffectiveConfigFile to read the settings misses the second appearence of the [status] section.

The user might end up with this .git/config if it manipulates it via "git config" and executes this sequence of valid git commands:

$ git config status.showUntrackedFiles no
$ git config --unset status.showUntrackedFiles
$ git config status.showUntrackedFiles no

This happened to me with "git version 1.9.5.msysgit.0	"