ToolStripGitStatus: Handle changes in the global ignore file.
The global ignore file isn't watched for changes, as described in #3054. This means that
when the file changes, we have to wait for the catchup timer to trigger (up to 5 minutes)
to have the commit strip updated.

Nothing fancy here, just a new FileWatcher and some plumbing to trigger a refresh when
needed.