feature: Allow auto-normalisation of branch name
fixes #3236

Git spec imposes a number of rules on how branches are named.
The rules can be found at https://www.git-scm.com/docs/git-check-ref-format/1.8.2

The feature allows to automatically check for compliance when the user
creates a new branch (by typing in or inserting from the clipboard).

The feature can be turned on/off at Settings->Advanced tab.
By default the feature is on, the replacement symbol is set to '_'.

The FormCreateBranch UI is reworked to provide a toggle for the option.