Load remote branches with FormRemoteProcess. Closes #573.
When branches are loaded from the remote, the ssh connection has to be established.
FormRemoteProcess is used to enable the user to input a password when openssh is chosen as a ssh client.
There is a option introduced to chose whether branch names should be read from the git's local database
or from the remote directly. The local database is set as a default.