refactor: Management of remote repositories
Remote repositories are accessible on a several forms, however the experience is quite inconsistent.
The management of the remotes is quite poor too, UX is confusing and buggy.
In the original implementation the UI and the BL concerns intertwined which makes it difficult to maintain, extend and test.

* The new implementation moves the functionality concerning dealing with the remotes repositories into a
separate controller abstracting the data management from the UI.
* The controller enables testing
* The controller also simplifies the remotes management (FormRemotes). The "remote repositories" tab's layout has also been reworked and updated.
* All affected forms now need to instantiate a controller, load the remotes and bind them to the desired UI control (i.e. to a combo).

The following reported bugs are fixed:
- #2550 - Remote Repositories > Separate Push URL not working (https://github.com/gitextensions/gitextensions/issues/2550)
- #1465 - Tab Stop Order in Remote Repositories dialog is very messed up (https://github.com/gitextensions/gitextensions/issues/1465)

The following change was undone, as it introduced quite inconsistent and confusing UX:
- #1208 - Rearrange buttons in "Manage Remote Repositories" dialog (https://github.com/gitextensions/gitextensions/pull/1208)

Translations:
The following items added in FormRemotes:
- lblMgtPuttyPanelHeader.Text
- _gitMessage.Text
- _gbMgtPanelHeaderNew.Text
- _gbMgtPanelHeaderEdit.Text
The following items removed from FormRemotes:
- buttonClose.Text
- PuTTYSSH.Text
- _hintDelete.Text
- groupBox1.Text

Fixed few Russian translations.