Fix "Host Fingerprint not registered" Plink handling on clone
When using Plink and pulling or pushing to/from a remote with an untrusted host fingerprint GitExtensions offers to open plink in a new command prompt window so the user can interactively press Y to trust the fingerprint.

However, this currently does not work correctly for clone. plink is called with an emptry string instead of the remote URL. This commit fixes that.