GitCommands: Avoid creating a fake remote ref on pull
This was done in eb3fce806cb76d328b669adf8de9a04244436172.

Passing +refs/heads/remote-branch:refs/remotes/origin/local-branch is
plain wrong. It creates a local ref for a remote branch which doesn't
really exist (origin/local-branch).

We don't really need a persistent ref for the remote branch. Just
passing it as a raw argument is fine. It is fetched into FETCH_HEAD,
then the local branch is merged/rebased on top of it.