feat: Dashboard revisited
Dashboard completely overhauled to provide simpler UX.
The dashboard can have two distinct themes - light and dark. When the
user uses the new dashboard for the very first time she will be presented
with a dialog prompt to pick the default theme.

The dashboard recent repositories list is sync'ed with the list of
repositories dropdown to provide the consistent UX.
The list is managed via the FormRecentReposSettings.

Also:
- DashboardCategory deprecated and removed
- DashboardItem deprecated and removed
- DashboardEditor deprecated and removed
- Repository category types removed
- RSS feeds deprecated and removed

#3658