FormOpenDirectory: Add a button to go (easily) to the parent directory
Most of the time the path selected when opening the form is the one of the current open repository.
It could be very convenient to be able to go to the parent directory that has a high probability to contains
a lot of git repositories (and open the "AutoComplete" list that show the available folders to select one of them easily).