fix: Pull dialog help picture disappear
Help images for helpImageDisplayUserControl1 are set whenever merge
options are changed - e.g. merge/rebase/don't merge.
The designer defaults to the first option (merge) after that the user
settings are applied.
If the user settings were set to 'merge' as well the change event would
not be triggered and images would not be set.

The designer default is unset, now the change event is triggered as expected.

Fixes #3829