fix: Help in Branch dialog exception
In a rare case when the user's browser of choice points to non-existent
executable attempts to launch the help will result in Win32Exception "the system
cannot find the file specified".
The old implementation was also looking at the now-defunct registry key to
determine the browser of choice.
Since Win8 the registry key containing the user's choice is encrypted and thus
no longer usable by thrid-party applications.

The code is simplified to launch the help URL via `Process.Start`.

Fixes #3011