GitModule is not updated for RevisionFileTreeController.cs
This is a regression for #4071. _module is not updated when the repo in FormBrowse/RevisionFileTree is updated as the value is passed at creation only
The handling is adapted to pass the module at updates

This has very minor negative effect: The icon provider cannot find the files on the file system so it creates a file in a temp dir
However, I wasted some time on RevisionDiffController trying to use the same pattern.