Always show artificial commits in RevisionGrid (#4147)
* Always show artificial commits in RevisionGrid

Also limit count of changes to the Browse window (does not really make sense in FileHistory)
This is #4031 item 2., discussed in #4086

The RevisionGrid is only updated at explicit commands as normal commits are not changed.
Artificial commits are dynamic and will have contents/no contents from many commands. However, they always "exist".
Now you have to refresh the the complete revision grid to see when staged or unstaged appear. As the diff lists are not cached for artificial commands, the content will be updated just by selecting the artificial commit in question.
(The count is not updated dynamically yet).

* Always have the Staged artificial commit as a parent to Unstaged