RevisionFileTree context menu gave exceptions if no items were Selected
Similar to #4098, related to #4031

Most easily seen by right clicking in the FileTree and selecting
The core change is in FileTreeContextMenu_Opening, disabling irrelevant items
Correction for open folder, clean to handle null selections too

The pattern to check for no selection was broken in most of the file. This was changed too. Probably not required after the menu change, could have been separated but easily reviewed.