fix: 'Unable to remove the file to be replaced' exception
Whenever <repo>/.gitconfig file gets updated there is a chance that the update would fail.
The issue is intermittent and may only happen on I/O-incapacitated systems
(i.e. virtualised environment with McAfee AV).

The problem seems to be caused by `File.Replace` API. The code is refactored not to use it.

Fixes #4250