Jenkins multibranch pipeline support, cleanup (#4279)
* Jenkins multibranch pipeline support, cleanup

Implements #4202
This is rewrite of how the plugin works
Previously the builds in a job were fetched first, then each individual build. Due to the implementation, the same info were retrieved many times. The stability were not good, not showing information often for bigger setups.
This has been changed to retrieve all builds with one request. If there are InProgress builds, all builds are fetched again, otherwise there is a check if there are new builds started on the server.

If there are many builds for one commit, the latest build is presented.