Git treefiles were not set as tracked, stash diff was inverted
Part of #4373
GitItemStatus.IsTracked should be set for all file-revisions handled by Git, i.e. all but working file directory files
This was not set for GetTreeFiles(). This is called for revisions without parents: First/detached commit and stash-untracked.
These files were therefore presented as the difference to the local file if local exists.

In addition, stash-diff was displayed inverted. This was a regression from #4157