Set VS2017 as default
Issue #4311
Building in AppVeyor with VS2017
This aligns the use of 2015/2017 as well as future versions and therefore removing MS2015 in filenames.
The known breaking future compatibility issues are new versions of .sln format as well as msbuild location (handled in hMSBuild.bat right now).

Also removed seemingly unused scripts and descriptions referring to MS2013 and older.