fix: Remotes not persisted
Since 2.50 user remotes are no longer persisted in the config.
The regression was introduced during the refactor of the FormRemotes
in 613335552bf0735080e442629da1b71b99551d53.

Restore the ability to save new or changed remotes by ensuring they are
added to the collection whenever 'Save' button is clicked.
This will allow saving remotes for new and currently configured remotes.

Deletion of remotes is still absent and requires further thoughts.
Deleting a remote from the history list may be undesirable as it would hinder
user's ability to *quickly* clone the remote repository.
The flipside is that the history list may grow long without a UI to manage it.

Fixes #4441

(cherry picked from commit 29aa23caadd4e2d23167647b8c5666465bb34890)