fix: Unable to solution build in VS2017
Unable to build current master in VS2017 on a computer which has only VS2017 installed.
It looks like a breaking change in VS2017 - a large number of Visual Studio assemblies are
no-longer installed in the GAC and which, in absence of previous versions of VS installed,
must be referenced in a different manner.
Further reading: https://developercommunity.visualstudio.com/content/problem/15434/cannot-find-wrapper-assembly-for-type-library-envd.html

Fixes: #4458