refactor: Extract IsArtificial functionality from GitRevision
GitRevision is a DTO yet it contains behavioural components.

Check for artificial commit is not a function of GitRevision.
Any revision inherently knows whether it is artificial or not. However checking an arbitrary hash for being artificial
is not a function of GitRevision.
For now, the functionality moved into an extension method. This way it can be tested and used universally.