Difference for initial commit of a file were presented as a diff to the working directory (#4585)
Fixes #4580
Regression from #4313

Non-existing commits are handled with the "revision" null which is mostly handled as a working directory file.
The compare options for working directory files were limited in 2.50. In 2.51 the changes are presented with git-diff if possible and the non-existing parent were incorrectly handled.
Also some async/await changes

Don't indicate "in progress" async actions with WaitCursor.
If needed, the "awaiting" status should be displayed by the control that is awaiting the async action.
This should be done only if the load time is significant.

Cherry-pick of 7f7b26499bce424864ada3c05d15bd9ab6534581