Annotate all StringExtensions and BoolExtensions methods
These annotations are used by ReSharper to find potential NREs and
impossible code paths. They also serve as documentation to developers
who do not use R#.

There's a proposal in place to add support for the nullability annotations
to the C# type system in future. If that happens, we'll be able to use
these annotations to drive that process.