Browse Diff tab menu alternatives and presentation (#4569)
* Browse Diff: Show differences for several selections
* Browse Diff tab menu alternatives and presentation

Fixes #4564
Fixes #4387
Fixes #4396

Depends on #4562 #4663 #4365 #4366 #4367 #4368

Rewrote RevisionDiffController to make it possible to have some kind of tests. The tests are retrofitted to the current functionality as they are just testing the menu items and not the actions themselves. (The tests adds may more maintenance than they give benefits but the formal test coverage increases.)

Browse Diff:
Submodules actions available for multi select (#4568)
Presents the Reset options for the correct parent in multi select situations
No longer differ between parent-child and first-second for Reset
Difftool: Describes Selected (B) and First (A) revisions in the menu
Difftool arguments depends on parent availability
Better detection of parents to A/B for diffs

Commit:
Limit DiffTool and FileHistory to tracked

FormDiff:
Limit FileHistory to Tracked
Use RevisionDiffController to align to Browse-Diff

* Extract common revision diff context menu logic into IRevisionDiffContextMenuController