Submodule method review (#4639)
* Update GitModule.GetSubmodulesLocalPaths XML documentation

* Refactor common code into nested function

* Make GitModule.GetSubmoduleInfo a nested function

* Rename local function to indicate it performs parsing

* Document output format of git submodule status

* Make GitSubmoduleInfo immutable

This commit also forces the name & remote path values to be populated
during response parsing. Previously the config file was loaded 2*N times,
now it's loaded once.

Parsing is also done using Regex meaning no string manipulation,
trimming, etc.

* Remove unused setters from interface

* Remove unused equality/hashing methods

Apart from the fact they're unused, they only test for partial equality,
and if needed would be better suited to a custom equality comparer.

* Remove unused GitModule.GetSubmoduleRemotePath

After the last commit, this method is no longer needed

* Don't read the .gitmodules file once per module

This commit hoists the read operation to before the loop.
The now-unused GitModule.GetSubmoduleNameByPath is removed.

* Validate in one place only

Old code attempted to skip lines that didn't validate, where
that validation was just a length check. This change
causes the code to skip lines that don't match the regex.

* Use operator for string equality test

* Rename locals

* Fix invalid assert

Probably meant to use Assert(false), but Fail is more clear.

* Don't re-use local with different meaning

* Assert submodule working dir has expected prefix

* Remove misleading #region

These members are not in IGitCommands.
They are in IGitModule.
However so are many other members in this class, scattered amongst
non-interface members.