Mouse highlights blame commit on LHS too
Previously only when hovering the mouse on the RHS would all
lines from that commit be highlighted.

With this change, the mouse highlights all lines from the hovered
commit in both the left and right views, no matter where the mouse
is located.

It also avoids calling GitBlameCommit.ToString on every mouse move
event then comparing the full string. Instead it does a quick reference
comparison on the commit object, and only calls ToString if needed.