Cache recently used avatar images in memory
Before this, caching always occurred on disk.

In-memory caching reduces disk IO during navigation, and
improves perceived responsiveness in the UI.

The assumption is that for many repos, there is a moderate number
of users committing in a given time span, and so caching their images
in memory is likely to yield a good hit rate and justify the memory
overhead.

This implementation decorates the existing file system cache.

Future work could make the file system cache decorate the
image download code. Such a pipeline would likely be the
simplest and cleanest solution.