Update the Plugins menu after plugins are loaded
Prior to this change, the Plugins menu was populated when it was first opened.
If the plugins were not already loaded at that time, the UI would latch in a
state where no options were available to users. This change does not prevent
the user from opening the Plugins menu before plugins are loaded, but it does
ensure the UI will not latch in a bad state if this occurs.