Improve RevisionGraph log parsing
Prior code used a more complex approach and did a lot more string
allocation. This commit introduces a regex, and modifies the git
log format so that there's only one \0 character, and it separates
log entries.

We also know the dates are numeric, so can parse confidently
(without TryParse).

The file name code was unused and has been removed.

The state machine is no longer needed as each log item
is separated by a \0 and can therefore be processed in one
go via the regex.