Reduce event args allocation
RevisionGrid fires this event once for every revision in the graph.
For the GE repo, this means ~11k useless object allocations
every time we refresh the revision grid.

I know there's a preference for using the "event args" pattern,
but in this case it's just pure waste.

A proper fix would be to not use events here. As I continue
working on these classes that may fall out.