Migrate legacy categorised user's git repos
Provides the ability to persist and load the collections of
categorised user's git repositories in the new format.
Also migrate the existing categorised git repositories from old format
into the new format.
Once migrated the old setting is supposed to be reset, however right now
the value is retained. This functionality will be enabled in the future.

Part of #4899