fix: Init wrong folder
When launched from Windows explorer, the clone destination was set to the parent of
the current folder. If the folder was at the root of the drive, the destination would be blank.

The fix sets the parent of the current directory unless the current directory is the root, or
the current directory is not a git repository.

Fixes #4855

(cherry picked from commit ae2f871026a80023d451e62b8a8c5b73bcbcec9a)