fix: Crash when you click on the Commit tab
Some users to the west of GM experience AOORE when navigating to
the artifical commits.

Artificial commits are constructed using DateTime.MaxValue, and for users
with timezones to the west of GM the MaxValue falls outside
DateTimeOffset.MaxValue.

The DateTimeOffset constructor first converts any DateTime that is not
of Kind 'UTC' to the equivalent UTC time. It will then check whether the
UTC-equivalent DateTime falls outside of the bounds of DateTimeOffset.MinValue
and DateTimeOffset.MaxValue, and if it does, will throw an
ArgumentOutOfRangeException.

Fixes #4956