Handle start with dashboard explicitly
Starting with dashboard is done if starting without parameters and AppSettings.StartWithRecentWorkingDir is not set. After the parsing of parameters, this was handled as start if FormBrowse had no working directory.
This implicit handling broke in #4740(#4860) and #4870 (tried to always resolve the current working directory when starting GE for other reasons).