Refactor JumpList and thumbnail toolbars management
Move the code related to Windows jumplists and thumbnail toolbars out of
FormBrowse into own set of classes.
The underlying functionality remains for the most part the same.

For the ease of use added an abstraction that represents a thumbnail
toolbar button; and a class that aggregate all available buttons.

Relates to #4549.