Remove 'Force Delete' functionality
The functionality adds a nice touch, however it can be extremely harmful,
if used carelessly. It also introduces an inconsistent UX, because other
parts of application do no allow deletion of refs without prior confirmation.

The functionality can be re-instated with an introduction of a user
controller setting (explicit opt in, placed on the Advanced tab), that
would enable force deletion of refs unversally all across the app.
This includes branch and tag deletion dialogs.

Relates to #4814