Revision grid columns don't stretch to fill space
Instead they have a given width, and a horizontal scrollbar is used if
needed.

By default DataGridView doesn't render the row background full width,
so some work was done to achieve that.

By default DataGridView doesn't hit test rows if you click in the 'empty'
space to the right of the last column. Some work was done to achieve
that also.

Widths and colours were updated.