fix: Disproportional controls' margin scaling
It looks like a bug in WinForms - NumericUpDown and TextBox controls'
margins seems to get scaled beyond expectations when Windows is set to
scale.
For example, if the design was done at 100% and Windows is set to scale
to 200%, NUD's margins get set to (96, 96, 96, 96) and TB's margins get
set to (12, 12, 12, 12).

Fixes #5098