fix: Current folder is not shown in FormInit
When attempting to init a repo in a folder the current folder woud not to be
shown to the user. The most recent repository was shown instead.

The issue is observed in databound lists, `SelectedIndex` must be reset before
setting an arbitrary value to `Text` property.

Fixes #5107

(cherry picked from commit 12adb464cb852e27c1fbce15b7de698497a4fe61)