fix: Don't load settings if not a valid git repository
If attempt to clone into a folder that is not a valid git folder,
but which contains 'config' file with an invalid (from GE's POV)
content, GE would attempt to load and parse the 'config' file
and crash the app.
This is a minimal fix to top this from happening by checking the
location for being a valid git repostitory.