Debug config uses release mode ICSharp.TextEditor
The debug configuration of that project includes a DATACONSISTENCYTEST
symbol that causes very slow assertion logic to run.

We used to take this dependency via NuGet, using release builds.

This change means we use the release build even during debugging,
as we would have before making the project a source submodule.