Guarantee several properties not null
Before this change, the following properties were annotated as
[CanBeNull], which caused R# to display warnings in the many places
that didn't check for null values.

- GitModuleForm.Module (402 usages)
- GitModuleControl.UICommands (66 usages)
- GitModuleControl.UICommandsSource (6 usages)
- GitModuleControl.Module (133 usages)
- GitUICommands.Module (62 usages)
- IGitUICommands.GitModule (19 usages)

Only five (0.7%) usages defended against null.

If you can't beat em, join em.

This change ensures these properties are not null. If somehow an
access occurs when they would be null, an InvalidOperationException is
thrown.