Convert static AsyncLoader.DoAsync to JTF
Previous implementation used a new instance of AsyncLoader for each call,
however this does not utilise the features of AsyncLoader and so it's simpler
and clearer to just use JTF directly here.

The next step will be to inline the few usages of this static form throughout the
code for better uniformity.