fix: Clone from FormBrowse does not prompt to open
`Application.OpenForms.Count` intermittently is 0, thus causing certain
forms to be shown in the taskbar and breaking certain codepaths.
For example, whenever `FormClone` is invoked and a new repo is cloned
the user must be prompted to open the new repository.
This prompt is driven by a check to see whether the `FormClone` is shown
in the taskbar or not.

There are also suggestions the observed behaviour is caused by a
WinForms' bug https://stackoverflow.com/a/3751748/2338036.

Fixes #5274